requirements.txt
Django==5.1.7
django-jazzmin==3.0.1
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
weasyprint==65.1
